Architecture: new class mvsData::imageAlgo to handle image processing
Move image processing functions from mvsData::imageIO to the new class mvsData::imageAlgo -> imageIO handles images reading/writing only
