+++ src/aliceVision/mvsUtils/fileIO.cpp
@@ -8,7 +8,7 @@
 #include <aliceVision/system/Logger.hpp>
 #include <aliceVision/mvsUtils/common.hpp>
 #include <aliceVision/mvsUtils/MultiViewParams.hpp>
-#include <aliceVision/mvsData/imageIO.hpp>
+#include <aliceVision/mvsData/imageAlgo.hpp>
 #include <aliceVision/mvsData/Image.hpp>
 
 #include <boost/filesystem/operations.hpp>
@@ -347,21 +347,21 @@ void loadImage(const std::string& path, const MultiViewParams* mp, int camId, Im
     {
         ALICEVISION_LOG_DEBUG("Downscale (x" << processScale << ") image: " << mp->getViewId(camId) << ".");
         Image bmpr;
-        imageIO::resizeImage(processScale, img, bmpr);
+        imageAlgo::resizeImage(processScale, img, bmpr);
         img.swap(bmpr);
     }
 
     if(bandType == 1)
     {
         Image smooth;
-        imageIO::convolveImage(img, smooth, "gaussian", 11.0f, 11.0f);
+        imageAlgo::convolveImage(img, smooth, "gaussian", 11.0f, 11.0f);
         img.swap(smooth);
     }
 
     if(bandType == 2)
     {
         Image bmps;
-        imageIO::convolveImage(img, bmps, "gaussian", 11.0f, 11.0f);
+        imageAlgo::convolveImage(img, bmps, "gaussian", 11.0f, 11.0f);
 
         for(int y = 0; y < height; y++)
         {
